Moved and renamed MySQLQuery class. Ripple changes in Repository class.
